removed redundant gl utility
